sysutils/DTraceToolkit: remove missing providers in FreeBSD 12.0-CURRRENT
Summary:
dtruss fails as syscall::lstat*:return is no longer present.

This may well be a problem in FreeBSD src that needs addressing, but
might as well start with a working port for the moment.

dtruss is a super handy sysadmin-level trace tool that would be a shame
to lose.

Reviewers: jrm, dteske, markj

Reviewed By: jrm, markj

Subscribers: mat

Differential Revision: https://reviews.freebsd.org/D15879